backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain ecosystem (compatible set)
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3
langchain-anthropic==0.1.6
langchain-groq==0.1.4

# Vector DB
chromadb==0.4.24
faiss-cpu==1.13.2

# Embeddings
sentence-transformers==2.3.1
openai==1.10.0

# Document processing
pypdf==3.17.4
unstructured==0.11.8
beautifulsoup4==4.12.3
python-docx==1.1.0

# Utilities
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
tenacity==8.2.3
tiktoken==0.5.2

# Streaming
sse-starlette==1.8.2

# Evaluation
rouge-score==0.1.2
nltk==3.8.1

typing-extensions==4.9.0